Initial claude rework
